Refactor layout interactions into separate module to fix dropdown initialization
