+++ sbomify/apps/core/js/htmx-bundle.ts
@@ -1,3 +1,2 @@
-import 'bootstrap';
-
+import './layout-interactions';
 import './alerts-global';
