+++ sbomify/apps/documents/js/main.ts
@@ -1,4 +1,5 @@
 import 'vite/modulepreload-polyfill'
+import '../../core/js/layout-interactions'
 
 import mountVueComponent from '../../core/js/common_vue'
 import DocumentUpload from './components/DocumentUpload.vue'
