+++ sbomify/apps/vulnerability_scanning/js/main.ts
@@ -1,4 +1,5 @@
 import 'vite/modulepreload-polyfill';
+import '../../core/js/layout-interactions';
 import { createApp, type Component } from 'vue';
 
 import VulnerabilityTimeSeries from './components/VulnerabilityTimeSeries.vue';
