+++ sbomify/static/css/layout.css
@@ -650,16 +650,48 @@ body {
   justify-content: center;
   width: 2.5rem;
   height: 2.5rem;
+  padding: 0 !important;
   text-decoration: none;
   color: var(--text-secondary);
   border-radius: 0.25rem;
   transition: all var(--transition-sidebar);
+  line-height: 1;
+  text-align: center;
+  position: relative;
+}
+
+.nav-icon i {
+  position: absolute;
+  top: 50%;
+  left: 50%;
+  transform: translate(-50%, -50%);
+  margin: 0;
+  line-height: 1;
+  width: 100%;
+  height: 100%;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+}
+
+.nav-icon.dropdown-toggle::after {
+  display: none;
 }
 
 .nav-icon:hover {
   background-color: rgba(var(--black-rgb), 0.05);
   color: var(--text-primary);
   text-decoration: none;
+}
+
+/* Center icon/text content inside workspace & settings toggles */
+.workspace-selector .nav-link,
+.workspace-selector .nav-icon,
+.nav-item.dropdown > .nav-link.dropdown-toggle {
+  display: inline-flex;
+  align-items: center;
+  justify-content: center;
+  text-align: center;
 }
 
 .dropdown-menu {
